Change JuliaSyntaxHighlighting UUID (#56852)
TLDR; I made a silly mistake a while back, and this seems like the least
bad option.

-----

Unfortunately during development of the backported version of this
stdlib, to avoid potential sysimage issues the UUID was temporarilly
changed and I forgot to change it back.

This mistake lead to the backport/compat verison of the package having a
different UUID from the in-tree stdlib. Beyond this being an embarassing
mistake, it's also a pain because it means that we now have
incorrect/invalid manifests.

A small mercy is that 1.12 hasn't been released yet, so the discrepancy
currently only affects people using development Julia versions. Changing
the UUID here will require these people to regenerate their manifests,
but this seems like the lesser of two evils.

Appologies for the hassle all.

Co-authored-by: Chengyu Han <cyhan.dev@outlook.com>