staticdata: Insert backedges recursively
In the new edges scheme, when we perform constant propagation, we create a dummy CodeInstance
that is not inserted into the specializations cache, but instead simply serves as a container
for all edges that were encountered during the constant propagation. These CodeInstances are
not part of either the internal or external edges list collected during pkgimage generation.
As such, while we were verifying edges recursively and would thus catch an invalidation
prior to package image reload, we were failing to also insert backedges for these CodeInstance.
We were thus failing to invalidate such methods if the method was redefined after re-load
of the package image. Fix that by moving the storing of the backedges to the end of the
validation code, so that it too happens recursively.