contrib: add `--relative-rpath` option to `juliac.jl` (#57199)
This flag is useful when you'd like to deploy the generated library /
executable across machines. This will tell the linker to look for the
required libraries (`libjulia*.so`, `libopenblas`, etc.) in an adjacent
`julia/` directory.

We'll eventually use PackageCompiler to help bundle all the required
libraries together, but for now this at least gets the lookup setup
correctly in the created `.so` / binary.

---------

Co-authored-by: Gabriel Baraldi <baraldigabriel@gmail.com>