Desugaring and runtime support for generated functions
I've opted to diverge from Base's lowering of `@ generated` code
generators to add a semi-hidden `__context__` argument to the lowered
code generator function in exact analogy to the context added to macros.
This makes the code generator even more macro-like, and I've also reused
the MacroExpansionContext.

We need our own `GeneratedFunctionStub` here in order to construct the
context, to call back into JuliaLowering's version of the lowering
machinery in a precise way, and to propagate enhanced provenance
information.

This all turned out to be quite subtle, but JuliaLowering is able to
almost fully integrate with the runtime without any changes to the
runtime itself. Thus we can see the effort people have put into building
abstractions for Cassette/etc has really paid off.