Actually compute sparams in jl_fptr_sparam
I was running a debug build of julia through the cmdlineargs tests
and saw it repeatedly hit the `sparams != jl_emptysvec` assertion.
That assertion doesn't make sense to me, since jl_fptr_sparam is
specifically used in the case where the sparams need to be recomputed:

https://github.com/JuliaLang/julia/blob/12698af99f2502227aba611f0107b1dc3a6872d2/src/codegen.cpp#L3038-L3043

It's possible I'm missing some larger point, but this fixes the
assertion.