Allow macrocall as function sig
`@KristofferC` requested that `function @main(args) end` should work.
This is currently a parse error. This PR makes it work as expected
by allowing macrocall as a valid signature in function (needs
to exapand to a call expr). Note that this is only the flisp changes.
The same changes were previously applied in JuliaSyntax.