Allow macrocall as function sig (#55040)
@KristofferC requested that `function @main(args) end` should work. This
is currently a parse error. This PR makes it work as expected by
allowing macrocall as a valid signature in function (needs to exapand to
a call expr). Note that this is only the flisp changes. If this PR is
accepted, an equivalent change would need to be made in JuliaSyntax.

(cherry picked from commit b65f0040dd367f2da76f6ed4b01924a368fe50b8)