Allow macrocall as function sig (#55040)
@KristofferC requested that `function @main(args) end` should work. This
is currently a parse error. This PR makes it work as expected by
allowing macrocall as a valid signature in function (needs to exapand to
a call expr). Note that this is only the flisp changes. If this PR is
accepted, an equivalent change would need to be made in JuliaSyntax.