trimming: make sure to fail / warn on `Expr(:call, ...)` (#57342)
This dispatch should not be treated as resolved just because arg0 is
constant. With the previous code, that meant it was eligible for
last-minute call resolution, but call-resolution in codegen is now
forbidden so this needs to fail unilaterally.

(cherry picked from commit 0b74d1752e4a02f9078e95bc9893619b2e52a0b6)