gf: add `concrete_lim` parameter for addl. compilable method limit
The intention here is to distinguish between:
 - matched methods whose compilable specializations are in an O(1):1
   ratio with literal source code, so that the methods explored via
   this query are bounded by a fixed-multiple of source code size.
 - specializing methods, which may be in a many-to-one mapping from
   MethodInstance to literal source. These require a stricter limit
   when exploring the call-graph since they can often be expanded
   to arbitrarily many specializations.

The first type of methods are called "concrete" in this PR, although
that's open to review / feedback. A better name (which is correct
for a wider set of reasonable predicates) might be "bounded-
specialization functions" or "finitely-specializable functions".

In any case, the analogy is that an "instance" of a function is a
specialization, similar to how an "instance" of an abstract type is
a concrete type. As you might expect, concrete / finite-unions can be
split / explored more aggressively than unboundedly-abstract ones.

Just a draft / hack to start discussion: The exact predicate still
needs fixing and the implementation of `ml_sortmatches` needs to be
adapted to apply concrete_lim properly. It also might be worth caching
the "concrete-ness" query as a flag on the `Method` object itself.