Prohibit binding replacement in closed modules during precompile
This applies the existing prohibition against introducing new bindings
in a closed module to binding replacement as well (for the same reason -
the change won't be persisted after reload). It is pretty hard to
even reach this point, since `eval` into closed modules is already
prohibited and there's no surface syntax for cross-module declaration,
but it is technically reachable from lowered IR. Further, in the future
we may make all of these builtins, which would make it easier. Thus, be
consistent now and fully disallow binding replacement in closed modules
during precompile.