Prohibit `import`ing or `using` Main during incremental compilation
An upcoming optimization will skip most binding validation if no binding
replacement has taken place in (sysimage, pkgimage) modules. However, as
a special case, we would like to treat `Main` as a non-sysimage module
because the addition of new bindings in `Main` is common and we would
like this to not ruin the optimization. To make this legal, we have to
prohibit `import`ing or `using` any `Main` bindings in pkgimages. I don't
think anybody actually does this, particularly, since `Main` is not considered
loading during precompile (so you have to use the main binding via (Core|Base|).Main),
and I can't think of any good semantic reason to want to do this, but
regardless, it does add additional restrictions to `using`/`import`, so
I wanted to break it out into its own PR.