bpart: Track whether any binding replacement has happened in image modules
This implements the optimization proposed in #57426 by keeping track of whether
any bindings were replaced in image modules (excluding `Main` as facilitated by #57426).
In addition, we augment serialization to keep track of whether a method body contains
any GlobalRefs that point to a loaded (system or package) image. If both of these
flags are true, we can skip scanning the body of the method, since we know that we
neither need to add any additional backedges nor were any of the referenced bindings
invalidated. The performance impact on end-to-end load time is small, but measurable.
Overall `@time using ModelingToolkit` consistently improves about 5% using this PR.
However, I should note that using time is still about 40% slower than 1.11. This
is not necessarily an Apples-to-Apples comparison as there were substantial other
changes on 1.12 (as well as current load-time-tunings targeting older versions),
but I wanted to put the number context.