bpart: Also partition ->deprecated
This repeats the exercise in #57405. This is required for correctness,
because the ->deprecated flag also affects `using` resolution (since
it makes the tagged binding weaker for `using` purposes). That said,
in general our `->deprecated` semantics have been somewhat underspecified
with lots of `XXX` comments in the surrounding code. This tries to define
the semantics to give a depwarn on *access* (read or write) when:

1. Either the binding itself is deprecated; or
2. The implicit imports pass through a deprecated binding.

However, we do not give depwarns on access to bindings that were explicitly
imported (although we do give those warnings on the import) - the reasoning
being that it's the import that needs to be adjusted not the access.

Additionally, this PR moves into the direction of making the depwarn a
semantic part of the global access, by adjusting codegen and inference
appropriately.