Don't return null pointer when asking for the type of a declared global (#57447)
The `_DECLARED` partition kind used to be considered `guard`, but we now
consider it equivalent to an Any-typed `_GLOBAL` (but with weaker
redefinition properties). That said, its `->restriction` is NULL, so add
it to the list of bindings that should return `nothing` here (and thus
`Any` from the bulitin) to fix #57446.

(cherry picked from commit 0163991000ab83bfd9f6b5ece695c1e1e6138ea7)