reword eachslice docs to allow other slices objects (#57428)
The wording of `eachslice` was overly strict in enforcing a `Slices`
return type, instead of allowing other objects with similar behavior but
e.g. potentially additional behaviors. The existence of `AbstractSlices`
suggests this was actually intended at some stage but did not make it
into the docs.

AxisKeys.jl already returns something else and it seems not to cause any
problems, but it would be good to be able to do this without breaking
the documented contract of the method.

---------

Co-authored-by: Lilith Orion Hafner <lilithhafner@gmail.com>