Fix typo in printed message and clarify a printed hint (#57511)
The block of messages including a warning, two notes, and two hints
regarding extending type constructors without qualification contains one
typo. The last hint can also be clarified a bit, as it's currently
ambiguous as to where the qualification needs to occur; it only needs to
be qualified in the method definition.

There may be a better way to clarify the qualification bit than what I
have here, happy to amend further.