Ensure module of macro method is used for AST scope
This change allows macros to be extended from modules other than
`parentmodule(mac)` and have the scope of the AST be that of the module
the extension is written in.

This matches Julia's current semantics for macro extensions though it's
unclear how useful this actually is when different macro methods are
distinguished based only on the number of arguments.