[mmtk-julia] Allow building Julia with variations of the binding (#57176)
This PR enables building the mmtk-julia binding with different
configurations other than non-moving Immix. While it was already
possible to do so when building from source, this should enable using
the binaries generated via the latest version from BinaryBuilder
(https://github.com/JuliaPackaging/Yggdrasil/pull/10357).
I have also decoupled MMTk from the code, such that it should be *an
option* for a third-party heap. This should be set using
`WITH_THIRD_PARTY_GC=mmtk` (in `Make.user` or as an environment
variable).

The different configurations of MMTk can be achieved by setting the
variables below (set in `Make.user`, for example):

| Variable      |       |        |
|---------------|--------------|---------------|
| `MMTK_PLAN`     | Immix        | StickyImmix   |
| `MMTK_MOVING`   | 0            | 1             |
| `MMTK_BUILD`    | release      | debug         |

Note that the actual code to support building with moving and sticky
immix will be added in future PRs.
I've also added some preliminary documentation in `doc/gc.md` about
building Julia with MMTk, including a link to the binding and an small
FAQ on what to do if you break the build of Julia+MMTk.