cleanup old builtins
The `_apply_pure` function only has one user, and that user is unsound
anyways and should not be used, so replace that with equivalent
`_call_in_world_total` call and remove unnecessary definitions.

The awkward distinction between `invokelatest` and `_call_latest` has
not been relevant (and indeed causes performance issues) since kwfunc
was introduced in #47157.