staticdata: Refactor sysimage loading
Introduce `jl_image_buf_t` to represent the in-memory details of an unparsed
system image and clean-up several global variables and split loading logic
in staticdata.c so that we (almost) get everything we need from the sysimage
handle at once.

This allows sysimage loading to be separated into three phases:
  1. Lookup the raw sysimage buffer as a `jl_image_buf_t`
  2. For .so sysimages, parse the sysimage and initialize JIT targets, etc.
  3. Finally load the sysimage into a `jl_image_t`

Care was taken to preserve the existing behavior of calling `jl_set_sysimg_so`
to configure the sysimage before initializing Julia, although this is likely
to be next on the chopping block in a follow-up.