Remove offset field from GenericIOBuffer
This field was added in d7b9ac8 when the buffer's underlying data was changed
to Memory. However, I think it's not a good tradeoff. The benefit of the offset
field is that, when a buffer is constructed from a Vector which has previously
been pushfirst!'ed to, no data needs to be copied into the vector.
However, this is probably rare.
The downside is that it makes the implementation more complex, the struct larger,
and adds extra overhead to all operations on the IO.

In this commit, we instead copy the data internally in the buffer if constructed
from a Vector with a non-zero memory offset.