Make no-body `function` declaration implicitly `global` (#57562)
These were the intended semantics of #57311 (and matches what it used to
do in 1.11). Note however that this differs from the body-ful form,
which now always tries to extend. Fixes #57546.

Note that this implementation is slightly inefficient since it goes
through a binding replacement.
However, there's a change coming down the line which optimizes these
replacements.
I do think it should eventually be refactored to just create the binding
directly, but that's a little bit
of a larger change.

(cherry picked from commit 7fa0c13b934a86e205c95742fd974f6158651830)