bpart: Allow inference/codegen to merge multiple partitions
When inference and codegen queries the partitions of a binding, they
often only care about a subset of the information in the partition.
At the moment, we always truncate world ranges to the most precise
partition that contains the relevant query. However, we can instead
expand the world range to cover all partitions that are equivalent
for the given query. To give a concrete example, both inference and
codegen never care about the exported flag in a binding partition, so
we should not unnecessarily truncate a world range just because an
export was introduced.

Further, this commit lays the ground work to stop invalidating code
for these same kinds of transitions, although the actual logic to do
that will come in a separate PR.