`getproperty(::Type,f::Symbol)`: type assert when `f === :parameters`
It is common to access the `parameters` field of a `DataType` using
`getproperties`. However, when the first argument is not inferred as
`DataType`, the return type infers as `Any`, instead of as
`Core.SimpleVector`. I believe this causes much type instability and
invalidation.

The proper fix would probably be to introduce a getter function,
something like `_get_datatype_parameters(x::DataType) = x.parameters`,
then use the getter instead of directly using `getproperties`. For
`Compiler`, we may have to introduce a getter anyway, as
`Compiler.getproperty === Core.getfield`, but outside of the compiler
this seems like a good enough hacky fix.