docs: Make docs tests pass when run from Main (#57629)
This fixes `using Test; include("test/docs.jl")`. It appears the
prinitng of types in this context changed at some point to always
include the context module, rather than do so conditionally. I do not
know when this happened, but it appears to be there in 1.10 and 1.11 as
far as I can tell. Regardless, this adjusts the test to the current
behavior to make it easier to run individual failing tests in the REPL.