inference: avoid creating `PartialStruct` that never exists at runtime
Specifically, when `form_partially_defined_struct` tries to create a
`PartialStruct`, this commit adds extra checks to prevent creating it
if it doesn't exist at runtime.
While it would be better to avoid propagating these runtime-invalid
object types altogether, that would require a major overhaul of the
dispatch system. As a fix within the current dispatch system, this
commit is proposed.

- closes JuliaLang/julia#57673

Co-authored-by: Cédric Belmant <cedric.bel@hotmail.fr>