inference: add internal SOURCE_MODE_GET_SOURCE mode
We were previously using typeinf_code for this, but that has slightly
different semantics for recursion detection and handling of some edge
cases related to caching (e.g. seen in #57634). Overall, we would like
to avoid those sorts of spurious differences in compilation. Instead
(re)introduce a `source_mode` called `SOURCE_MODE_GET_SOURCE` which can
be used internally (requires a NativeInterpreter) here to avoid needing
to duplicate the existing `typeinf_ext` code to implement this use case.

While we are at it, also remove the code duplication for `typeinf_type`.